type of PathPoint's x and y field is double
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@14529 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/euclidian/PathPoint.java
@@ -2,7 +2,7 @@ package geogebra.common.euclidian;
 
 public abstract class PathPoint extends geogebra.common.awt.Point2D {
 	
-	protected int x;
-	protected int y;
+	protected double x;
+	protected double y;
 
 }
